Add PULSE docs
