--- examples/pulse/README.md
@@ -1,46 +1,0 @@
-# pi-topPULSE examples
-
-## cputemp.py
-Displays the CPU temperature in celsius on the pi-topPULSE's LED matrix.
-
-* Kindly provided by Rene Richarz
-* Based on [the script to display digits on the sense hat by yaab-arduino.blogspot.ch](http://yaab-arduino.blogspot.co.uk/2016/08/display-two-digits-numbers-on-raspberry.html)
-
-## led-demo.py
-Demonstrates multiple pattern effects that can be produced on a pi-topPULSE.
-
-* based on Unicorn Hat's LED demo code
-
-## led-pong-demo.py
-Demonstrates a pong game clone on a pi-topPULSE.
-
-## mic-demo.py
-Demonstrates capturing different audio formats to WAV files using the onboard microphone on a pi-topPULSE.
-
-## pt-pulse-battery
-Visually shows a pi-top's remaining battery level on the pi-topPULSE's LED matrix.
-
-![pi-topPULSE battery indicator](https://static.pi-top.com/images/pt-pulse-battery-small.jpg "pi-topPULSE battery indicator")
-
-* Kindly provided by Rene Richarz
-* Originally maintained [here](https://github.com/rricharz/pi-top-pulse-battery-gauge)
-* Based on [PiTop-Unicorn-Battery-Gauge](JezShed/PiTop-Unicorn-Battery-Gauge) by Jez Shed
-
-*NOTE: This script requires `pt-hub` to be installed to communicate with a pi-top hub's battery.*
-
-## showdigits.py
-Demonstrates how to show multiple pattern effects that can be produced on a pi-topPULSE.
-
-* Kindly provided by Rene Richarz
-
-## showusage.py
-Shows current CPU usage of the Raspberry Pi's 4 cores on the pi-topPULSE's LED matrix.
-
-* Kindly provided by Rene Richarz
-
-## testcolors.py
-Demonstrates colour representation of the pi-topPULSE's LED matrix by displaying 49 different colour intensities in the 49 pixels of the LED matrix.
-
-* Kindly provided by Rene Richarz
-
-*NOTE: Due to the way that the hardware works, each colour is 5 bits. Therefore, multiple pixels have the same intensity.*
