+++ examples/pulse/leds-cpu_temp.py
@@ -1,13 +1,14 @@
-#! /usr/bin/python3
+from pitop.pulse import ledmatrix
 
-# Display a 2 digit number on the pi-topPULSE led matrix
-#
-# based on the script for the sense hat by yaab-arduino.blogspot.ch
-# adapted for the pi-topPULSE by @rricharz
-#
-
-from ptpulse import ledmatrix
 import time
+
+
+def getCpuTemperature():
+    tempFile = open("/sys/class/thermal/thermal_zone0/temp")
+    cpu_temp = tempFile.read()
+    tempFile.close()
+    return int(int(cpu_temp) / 1000)
+
 
 OFFSET_LEFT = 0
 OFFSET_TOP = 2
@@ -44,16 +45,26 @@ def show_number(val, r, g, b):
     show_digit(units, OFFSET_LEFT+4, OFFSET_TOP, r, g, b)
 
 
-#############################################################
+###########################################################
 # MAIN
-#############################################################
+###########################################################
 
 ledmatrix.rotation(0)
 ledmatrix.clear()
 
-for i in range(0, 100):
-    show_number(i, 255, 255, 0)
-    time.sleep(0.5)
+lastTemperature = -1
+
+while True:
+    temperature = getCpuTemperature()
+    if temperature != lastTemperature:
+        if temperature < 60:
+            show_number(temperature, 0, 255, 0)
+        elif temperature < 70:
+            show_number(temperature, 255, 255, 0)
+        else:
+            show_number(temperature, 255, 0, 0)
+        lastemperature = temperature
+    time.sleep(2)
 
 ledmatrix.clear()
 ledmatrix.show()
