+++ examples/pulse/leds-test_colors.py
@@ -1,31 +1,14 @@
-#! /usr/bin/python3
-
-# Test color representation of the pi-topPULSE led matrix
-#
-# Displays 49 different color intensities in the 49 pixels of the lex matrix
-#
-# by @rricharz
-#
-# Note:
-# The hardware representation of each color is 5 bits e.g. only 32 different values
-# Without gamma correction, this would mean the actual color value
-# changes only every 8th color intensity value
-# The pi-top library also applies a gamma correction
-# Therefore multiple pixels have the same intensity
-#
-
-from ptpulse import ledmatrix
+from pitop.pulse import ledmatrix
 import time
 
 
-def showMap(r, g, b):
+def show_map(r, g, b):
     for x in range(0, 7):
         for y in range(0, 7):
             z = (float(y) + 7.0 * float(x)) / 49.0
             rr = int(z * r)
             gg = int(z * g)
             bb = int(z * b)
-            # print(x,y,rr,gg,bb)
             ledmatrix.set_pixel(x, y, rr, gg, bb)
     ledmatrix.show()
 
@@ -33,6 +16,7 @@ def showMap(r, g, b):
 ledmatrix.rotation(0)
 ledmatrix.clear()
 
+# Display 49 different color intensities
 for r in range(0, 2):
     for g in range(0, 2):
         for b in range(2):
@@ -41,7 +25,7 @@ for r in range(0, 2):
                 gg = 255 * g
                 bb = 255 * b
                 print(rr, gg, bb)
-                showMap(rr, gg, bb)
+                show_map(rr, gg, bb)
                 time.sleep(5)
 
 ledmatrix.clear()
