+++ pitopcli/oled.py
@@ -1,4 +1,5 @@
 #!/usr/bin/python3
+
 from time import sleep
 from os.path import isfile
